refactor: modernize tests
